+++ src/main/java/com/itlize/Korera/services/serviceImpl/ProjectServiceImpl.java
@@ -1,13 +1,18 @@
 package com.itlize.Korera.services.serviceImpl;
 
-import com.itlize.Korera.dbModels.Project;
-import com.itlize.Korera.dbModels.User;
+import com.itlize.Korera.dbModels.*;
 import com.itlize.Korera.repositories.ProjectRepository;
 import com.itlize.Korera.repositories.UserRepository;
+import com.itlize.Korera.services.ColumnsService;
 import com.itlize.Korera.services.ProjectService;
+import com.itlize.Korera.services.ProjectToResourceService;
+import com.itlize.Korera.services.ResourceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 
@@ -17,9 +22,19 @@ public class ProjectServiceImpl implements ProjectService {
     private ProjectRepository projectRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private ProjectToResourceService projectToResourceService;
+    @Autowired
+    private ResourceService resourceService;
+    @Autowired
+    private ColumnsService columnsService;
 
     @Override
     public boolean create(Project project, User user) {
+        if(project==null || user ==null) {
+            System.out.println("null input!");
+            return false;
+        }
         Project target = get(project.getId());
         if(target!=null){
             System.out.println("project already exists");
@@ -27,11 +42,13 @@ public class ProjectServiceImpl implements ProjectService {
         }
         try{
             projectRepository.save(project);
+
             user.addProjects(project);
             projectRepository.save(project);
             userRepository.save(user);
         }catch (Exception e){
             System.out.println("Sth wrong happens when creating: " + e.getMessage());
+            e.printStackTrace();
             return false;
         }
         return  true;
@@ -39,11 +56,20 @@ public class ProjectServiceImpl implements ProjectService {
 
 
     @Override
+    @Transactional
     public boolean delete(Project project) {
+        if(project==null){
+            System.out.println("null input");
+            return false;
+        }
+        System.out.println("deleting project: " +project.getId());
+
         try{
+
             projectRepository.delete(project);
         }catch (Exception e){
             System.out.println("Sth wrong happens when deleting");
+            e.printStackTrace();
             return false;
         }
         return  true;
@@ -58,5 +84,24 @@ public class ProjectServiceImpl implements ProjectService {
             return res.get();
         }
         return null;
+    }
+
+    @Override
+    public String toJson(Integer id) {
+        Project project = get(id);
+        if(id==null||project==null)
+            return "{\"error\":\"Project does not exist!\"}";
+        List<ProjectToResource> resources = projectToResourceService.get(project);
+        List<String> resourceJsons = new ArrayList<String>();
+        for(ProjectToResource ptr : resources){
+            String resourceJson = resourceService.toJson(ptr.getResource(),project);
+            resourceJsons.add(resourceJson);
+        }
+        return "[" + String.join("", resourceJsons) + "]";
+    }
+
+    @Override
+    public void clear() {
+        projectRepository.deleteAll();
     }
 }
