+++ src/main/java/com/itlize/Korera/services/serviceImpl/UserServiceImpl.java
@@ -1,17 +1,24 @@
 package com.itlize.Korera.services.serviceImpl;
 
+import com.itlize.Korera.dbModels.Project;
 import com.itlize.Korera.dbModels.User;
 import com.itlize.Korera.repositories.UserRepository;
+import com.itlize.Korera.services.ProjectService;
 import com.itlize.Korera.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
     @Autowired
     UserRepository userRepository;
+    @Autowired
+    ProjectService projectService;
 
     @Override
     public boolean create(User user) {
@@ -25,7 +32,13 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional
     public boolean delete(User user) {
+        if(user==null){
+            return false;
+        }
+        System.out.println("deleting user: " +user.getUserName());
+
         userRepository.delete(user);
         return true;
     }
@@ -53,5 +66,10 @@ public class UserServiceImpl implements UserService {
             return false;
         }
         return true;
+    }
+
+    @Override
+    public void clear() {
+        userRepository.deleteAll();
     }
 }
